Rename logger.js to log.js
